refactor: queryService CommandService 분리
