+++ src/main/java/nextstep/favorite/application/FavoriteQueryService.java
@@ -1,35 +1,30 @@
 package nextstep.favorite.application;
 
-import nextstep.favorite.payload.FavoriteRequest;
-import nextstep.favorite.payload.FavoriteResponse;
 import nextstep.favorite.domain.Favorite;
 import nextstep.favorite.domain.FavoriteRepository;
+import nextstep.favorite.payload.FavoriteRequest;
+import nextstep.favorite.payload.FavoriteResponse;
 import nextstep.station.domain.Station;
 import nextstep.station.payload.StationResponse;
 import nextstep.station.repository.StationRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-
+@Transactional(readOnly = true)
 @Service
-public class FavoriteService {
+public class FavoriteQueryService {
     private final FavoriteRepository favoriteRepository;
 
     private final StationRepository stationRepository;
 
-    public FavoriteService(final FavoriteRepository favoriteRepository, final StationRepository stationRepository) {
+    public FavoriteQueryService(final FavoriteRepository favoriteRepository, final StationRepository stationRepository) {
         this.favoriteRepository = favoriteRepository;
         this.stationRepository = stationRepository;
-    }
-
-    public Long createFavorite(final Long memberId, final FavoriteRequest request) {
-        Favorite favorite = new Favorite(memberId, request.getSource(), request.getTarget());
-        Favorite saved = favoriteRepository.save(favorite);
-        return saved.getId();
     }
 
 
@@ -56,12 +51,4 @@ public class FavoriteService {
                 .collect(Collectors.toMap(Station::getId, (station -> station)));
     }
 
-    /**
-     * TODO: 요구사항 설명에 맞게 수정합니다.
-     *
-     * @param id
-     */
-    public void deleteFavorite(Long id) {
-        favoriteRepository.deleteById(id);
-    }
 }
